Added utilities to retrieve last.fm song info.
To generate a large list of songs (may be useful later), last.fm's API
is used to retrieve a large list of users and then get information about
their favourite songs. last.fm's API allows anyone to retrieve this
data without authenticating themselves first.

First, the scrapy web-crawling tool was used to scrape the list of users
from last.fm's list of recently active users. However, since last.fm's
website was updated some time last year, this page is not accessible. So
instead a cached version of the page is used.

From this initial list of ~20 users, the pylast python module is used to
retrieve each user's friends. A breadth-first search is then performed
on the friends, providing virtually unlimited access to user names.
Using this method, approximately 400,000 user names were retrieved from
last.fm.

Then with this master list of user names, each user's top tracks are
queried and returned. Only the top 5 tracks are retrieved for each user
because of the large delay experienced otherwise. The top 5 "tags" for each of
these tracks are then retrieved as well. Tags are somewhat similar to
genres, and can be things such as "jazz", "guitar", "happy".

Once the list of tracks and tags is created, Echo Nest is used to get
information about each track. The standard audio-summary is requested
for each track and the summary is then added to the track's list entry.